no comment

+++ include/common.h
@@ -3,8 +3,10 @@
 
 #include <stdint.h>
 
-// TODO 根据平台选择
-typedef uint32_t addr_t;
+#include "include/list.h"
+
+// 根据平台判断，目前支持 ARM-Linux
+typedef uint64_t addr_t;
 typedef uint32_t instr_t;
 
 #endif
